+++ src/go/cmdinit/init_command.go
@@ -2,6 +2,8 @@ package cmdinit
 
 import (
 	"fmt"
+	"os"
+	"path/filepath"
 
 	"github.com/kkrull/marmot/cmd"
 	"github.com/kkrull/marmot/usemetarepo"
@@ -25,18 +27,24 @@ func (cliCmd *initCommand) toCobraCommand() *cobra.Command {
 	return &cobra.Command{
 		Long: "Initialize a new Meta Repo, if none is already present in the configured directory.",
 		RunE: func(cobraCmd *cobra.Command, _args []string) error {
-			metaHomeFlag := cobraCmd.Flags().Lookup("meta-home")
-			metaHomePath := metaHomeFlag.Value.String()
-			fmt.Printf("[init_command] meta-home: %s\n", metaHomePath)
-
-			if runErr := cliCmd.initAppCmd.Run(metaHomePath); runErr != nil {
-				return fmt.Errorf("failed to initialize meta repo at %s; %w", metaHomePath, runErr)
+			if metaRepoPath, pathErr := defaultMetaRepoPath(); pathErr != nil {
+				return pathErr
+			} else if runErr := cliCmd.initAppCmd.Run(metaRepoPath); runErr != nil {
+				return runErr
 			} else {
-				fmt.Printf("Initialized meta repo at %s\n", metaHomePath)
+				fmt.Printf("Initialized meta repo at %s\n", metaRepoPath)
 				return nil
 			}
 		},
 		Short: "Initialize a meta repo",
 		Use:   "init",
+	}
+}
+
+func defaultMetaRepoPath() (string, error) {
+	if homeDir, homeErr := os.UserHomeDir(); homeErr != nil {
+		return "", fmt.Errorf("failed to locate home directory; %w", homeErr)
+	} else {
+		return filepath.Join(homeDir, "meta"), nil
 	}
 }
